--- client/src/css/Scheduling/ScheduleView/CalendarTimeline.css
@@ -1,62 +1,0 @@
-:root {
-    /* colors */
-    --light-gray: #FAFAFA;
-    --gray: #D4D4D4;
-    --dark-gray: #D9D9D9;
-
-    /* fonts */
-    --title: 'Poppins', monospace;
-    --content: 'DM Sans', sans-serif;
-}
-
-.calendar-timeline {
-    max-height: 100%;
-}
-
-.fc-timegrid-slots tr {
-    height: 3rem;
-}
-
-.fc-theme-standard td, .fc-theme-standard th {
-    border-left: none;
-    border-right: none;
-}
-
-.fc-theme-standard .fc-scrollgrid {
-    border: none;
-}
-
-/* Date text element */
-.fc .fc-col-header-cell-cushion {
-    color: var(--gray);
-    font-family: var(--content);
-    background-color: var(--light-gray);
-    border-radius: 4%;
-    width: 40%;
-    text-align: center;
-}
-
-/* Time text element */
-.fc .fc-timegrid-slot-label {
-    vertical-align: bottom;
-}
-
-.event-content {
-    display: flex;
-    height: 100%;
-}
-
-.vertical {
-    border-left: 6px solid blue;
-    height: 92%;
-    margin-left: 2%;
-    position: absolute;
-    top: 3%;
-    left: 2%;
-}
-
-.event-description {
-    position: absolute;
-    left: 10%;
-}
-
