--- client/src/css/Scheduling/ScheduleView/ScheduleView.css
@@ -1,6 +1,0 @@
-.schedule-view {
-    width: 75%;
-    height: 80%;
-    margin-top: 10.7%;
-    margin-left: 23%;
-}
